# swagger.yml
produces:
  - application/json
consumes:
  - application/json
schemes:
  - http
  - https
swagger: "2.0"
info:
  title:  dex_api
  description: DEX API specifications.
  version: 0.1.0
paths:
  /api/projects:
    get:
      summary: Get list of projects.
      description: Get list of projects.
      operationId: listProjects
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/ProjectArray"
        "500":
          description: internal error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/firehoses/{firehoseUrn}:
    parameters:
      - in: path
        name: firehoseUrn
        type: string
        required: true
        description: URN of the firehose.
    get:
      summary: Get firehose by URN.
      operationId: getFirehose
      description: Get firehose by URN.
      responses:
        "200":
          description: Found firehose with given URN
          schema:
            $ref: "#/definitions/Firehose"
        "404":
          description: Firehose with given URN was not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: internal error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      summary: Update firehose configurations.
      description: Update firehose configurations.
      operationId: updateFirehose
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              description:
                type: string
                example: "This firehose consumes from booking events and ingests to redis"
              configs:
                type: object
                $ref: "#/definitions/FirehoseConfig"
      responses:
        "200":
          description: Found firehose with given URN
          schema:
            $ref: "#/definitions/Firehose"
        "400":
          description: Update request is not valid.
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Firehose with given URN was not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: internal error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  ErrorResponse:
    type: object
    properties:
      message:
        type: string
        example: "Request is invalid"
      cause:
        type: string
        example: "name must not be empty"
      code:
        type: string
        example: "internal_error"
        enum:
          - conflict
          - not_found
          - bad_request
          - internal_error
  ProjectArray:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/Project"
  Project:
    type: object
    properties:
      id:
        type: string
        example: "foo"
        readOnly: true
      name:
        type: string
        example: "Project Foo"
      slug:
        type: string
        example: "project-foo"
      created_at:
        type: string
        format: date-time
        example: "2022-06-23T16:49:15.885541Z"
        readOnly: true
      updated_at:
        type: string
        format: date-time
        example: "2022-06-23T16:49:15.885541Z"
        readOnly: true
      metadata:
        type: object
        example:
          environment: "production"
          landscape: "systems"
          organization: "foo"
...
  Firehose:
    type: object
    required:
      - configs
      - title
      - group
    properties:
      urn:
        type: string
        example: "orn:foo:firehose:fh1"
        readOnly: true
      title:
        type: string
        example: "Booking Events Ingester"
      name:
        type: string
        example: "booking-events-ingester"
        pattern: "^[A-Za-z][\\w-]+[A-Za-z0-9]$"
      project:
        type: string
        example: "g-goto-id"
      created_at:
        type: string
        format: date-time
        example: "2022-06-23T16:49:15.885541Z"
        readOnly: true
      updated_at:
        type: string
        format: date-time
        example: "2022-06-23T16:49:15.885541Z"
        readOnly: true
      description:
        type: string
        example: "This firehose consumes from booking events and ingests to redis"
      group:
        type: string
        format: uuid
        example: "e144ea5e-c7d6-48c4-a580-db31cb3389aa"
      configs:
        $ref: "#/definitions/FirehoseConfig"
      state:
        $ref: "#/definitions/FirehoseState"
      labels:
        type: object
        additionalProperties:
          type: string
  FirehoseConfig:
    type: object
    required:
      - stream_name
      - env_vars
      - kube_cluster
    properties:
      stopped:
        type: boolean
      image:
        type: string
        example: "gotocompany/firehose:0.1.0"
      stream_name:
        type: string
      kube_cluster:
        type: string
        example: "orn:entropy:kubernetes:sample_project:sample_name"
      replicas:
        type: number
        default: 1
        minimum: 1
      stop_time:
        type: string
        format: date-time
      deployment_id:
        type: string
      env_vars:
        type: object
        additionalProperties:
          type: string
  FirehoseState:
    type: object
    properties:
      status:
        type: string
      output:
        type: object
        additionalProperties: true
...